backend/requirements.txt
# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
websockets==12.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.31.0
alembic==1.13.1

# Redis / Background Tasks
redis==5.0.1
celery==5.3.6

# AI / Audio
openai==1.12.0
deepgram-sdk==3.0.0
edge-tts==6.1.9
pydub==0.25.1

# Utilities
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
httpx==0.26.0
